Refac: reorganize imports and modules in lightning
