+++ src/careamics/compat/model_io/bmz_io.py
@@ -19,7 +19,7 @@ from careamics.compat.config.utils.configuration_io import (
 )
 from careamics.compat.lightning.lightning_module import FCNModule
 from careamics.config.support import SupportedArchitecture
-from careamics.lightning.vae_lightning_module import VAEModule
+from careamics.lightning.modules.vae_lightning_module import VAEModule
 from careamics.utils.version import get_careamics_version
 
 from .bioimage import (
