+++ src/careamics/lightning/__init__.py
@@ -1,25 +1,34 @@
 """CAREamics PyTorch Lightning modules."""
 
 __all__ = [
+    "CAREModule",
+    "CareamicsCheckpointInfo",
+    "CareamicsDataModule",
     "DataStatsCallback",
     "MicroSplitDataModule",
+    "N2VModule",
     "PredictionStoppedException",
     "ProgressBarCallback",
     "StopPredictionCallback",
     "VAEModule",
+    "convert_prediction",
     "create_microsplit_predict_datamodule",
     "create_microsplit_train_datamodule",
 ]
 
 from .callbacks import (
+    CareamicsCheckpointInfo,
     DataStatsCallback,
     PredictionStoppedException,
     ProgressBarCallback,
     StopPredictionCallback,
 )
-from .microsplit_data_module import (
+from .data import CareamicsDataModule
+from .data.microsplit_data_module import (
     MicroSplitDataModule,
     create_microsplit_predict_datamodule,
     create_microsplit_train_datamodule,
 )
-from .vae_lightning_module import VAEModule
+from .modules import CAREModule, N2VModule
+from .modules.vae_lightning_module import VAEModule
+from .prediction import convert_prediction
