+++ src/careamics/lightning/data/__init__.py
@@ -1,0 +1,11 @@
+"""CAREamics Lightning Data Modules."""
+
+__all__ = [
+    "CareamicsDataModule",
+    "InputVar",
+]
+
+from .data_module import (
+    CareamicsDataModule,
+    InputVar,
+)
