+++ src/careamics/lightning/data/data_module.py
