+++ src/careamics/lightning/load_checkpoint.py
@@ -9,8 +9,8 @@ from careamics.config.configuration import AlgorithmConfig, Configuration
 from careamics.config.factories.config_discriminators import instantiate_config
 from careamics.config.support import SupportedAlgorithm
 
-from .lightning_modules import CAREModule, N2VModule
-from .lightning_modules.get_module import get_module_cls
+from .modules import CAREModule, N2VModule
+from .modules.get_module import get_module_cls
 
 CAREamicsModuleCls = type[N2VModule] | type[CAREModule]
 CAREamicsModule = N2VModule | CAREModule
