+++ src/careamics/lightning/modules/__init__.py
