+++ src/careamics/lightning/modules/get_module.py
