+++ src/careamics/lightning/modules/vae_lightning_module.py
