+++ src/careamics/metrics/psnr.py
