+++ tests/interoperability/test_lightning_module_onnx_exportability.py
@@ -3,7 +3,7 @@ import torch
 from onnx import checker
 
 from careamics.config.factories.algorithm_factory import algorithm_factory
-from careamics.lightning.lightning_modules import get_module_cls
+from careamics.lightning.modules import get_module_cls
 
 
 @pytest.mark.parametrize(
